+++ kite/src/main/kotlin/com/tang/kite/sql/dialect/OracleDialect.kt
@@ -1,26 +1,27 @@
 package com.tang.kite.sql.dialect
 
-import com.tang.kite.constants.SqlString.LIMIT
 import com.tang.kite.constants.SqlString.OFFSET
 import com.tang.kite.constants.SqlString.QUESTION_MARK
+import com.tang.kite.constants.SqlString.ROWS_FETCH_NEXT
+import com.tang.kite.constants.SqlString.ROWS_ONLY
 import com.tang.kite.sql.LimitClause
 import com.tang.kite.sql.enumeration.DatabaseType
 
 /**
- * PostgreSQL SQL provider
+ * Oracle SQL provider
  *
  * @author Tang
  */
-class PostgresqlDialect : SqlDialect {
+class OracleDialect : SqlDialect {
 
     override fun getType(): DatabaseType {
-        return DatabaseType.POSTGRE_SQL
+        return DatabaseType.ORACLE
     }
 
     override fun applyLimitClause(sql: StringBuilder, parameters: MutableList<Any?>, limitClause: LimitClause) {
-        parameters.add(limitClause.pageSize)
         parameters.add((limitClause.pageNumber - 1) * limitClause.pageSize)
-        sql.append("$LIMIT$QUESTION_MARK$OFFSET$QUESTION_MARK")
+        parameters.add(limitClause.pageSize)
+        sql.append("$OFFSET$QUESTION_MARK$ROWS_FETCH_NEXT$QUESTION_MARK$ROWS_ONLY")
     }
 
 }